+++ pkg/analyzer/time.go
@@ -88,17 +88,19 @@ func (t *Time) AvgTimeOnPage(filter *Filter) ([]model.TimeSpentStats, error) {
 	var query strings.Builder
 	t.selectAvgTimeSpentPeriod(filter.Period, &query)
 	fields := q.getFields()
+	filterFields := strings.Join(fields, ",")
+
+	if filterFields != "" {
+		filterFields = "," + filterFields
+	}
+
 	fields = append(fields, "duration_seconds")
-	query.WriteString(fmt.Sprintf(`SELECT "day",
-		toUInt64(ifNotFinite(round(avg(time_on_page)), 0)) average_time_spent_seconds
+	query.WriteString(fmt.Sprintf(`SELECT "day", toUInt64(ifNotFinite(round(avg(time_on_page)), 0)) average_time_spent_seconds
 		FROM (
-			SELECT "day",
-			%s time_on_page
+			SELECT "day", %s time_on_page, sid, neighbor(sid, 1, null) next_sid %s
 			FROM (
-				SELECT session_id,
-				toDate(time, '%s') "day",
-				%s
-				FROM page_view v `, t.analyzer.timeOnPageQuery(filter), filter.Timezone.String(), strings.Join(fields, ",")))
+				SELECT session_id sid, toDate(time, '%s') "day", "time", %s
+				FROM page_view v `, t.analyzer.timeOnPageQuery(filter), filterFields, filter.Timezone.String(), strings.Join(fields, ",")))
 
 	if len(filter.EntryPath) != 0 || len(filter.ExitPath) != 0 {
 		query.WriteString(fmt.Sprintf(`INNER JOIN (
@@ -114,11 +116,16 @@ func (t *Time) AvgTimeOnPage(filter *Filter) ([]model.TimeSpentStats, error) {
 		ON v.visitor_id = s.visitor_id AND v.session_id = s.session_id `, q.whereTime()[len("WHERE "):]))
 	}
 
-	query.WriteString(fmt.Sprintf(`WHERE %s ORDER BY visitor_id, session_id, time)
-		WHERE session_id = neighbor(session_id, 1, null) AND time_on_page > 0 `, q.whereTime()[len("WHERE "):]))
+	query.WriteString(fmt.Sprintf(`WHERE %s ORDER BY visitor_id, sid, time)
+		ORDER BY "time")
+		WHERE time_on_page > 0
+		AND sid = next_sid `, q.whereTime()[len("WHERE "):]))
 	q.whereFields()
 	where := q.q.String()
-	query.WriteString(fmt.Sprintf(`%s) GROUP BY "day" ORDER BY "day" %s`, where, q.withFill()))
+	query.WriteString(fmt.Sprintf(`%s
+		GROUP BY "day"
+		ORDER BY "day"
+		%s`, where, q.withFill()))
 	t.groupByPeriod(filter.Period, &query)
 	stats, err := t.store.SelectTimeSpentStats(filter.Period, query.String(), q.args...)