+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 6.5.2
+
+* improved tests related to time on page
+* fixed average time on page calculation
+* fixed events resetting time the session was last seen
+
 ## 6.5.1
 
 * improved tests related to session duration and time on page